python/samples/getting_started_with_agents/multi_agent_orchestration_lab/requirements.txt
chainlit>=1.0.0
azure-identity
semantic-kernel
